requirements.txt
pandas
numpy
scikit-learn
nltk
imbalanced-learn
pyyaml
wordcloud
matplotlib
seaborn
gdown
